requirements.txt
absl-py==0.9.0
aiofiles==0.5.0
aiohttp==3.6.2
alabaster==0.7.10
anaconda-client==1.6.9
anaconda-navigator==1.7.0
anaconda-project==0.8.2
aniso8601==8.0.0
appdirs==1.4.3
APScheduler==3.6.3
asn1crypto==0.24.0
astor==0.8.1
astroid==1.6.1
astropy==2.0.3
async-generator==1.10
async-timeout==3.0.1
attrs==19.3.0
Authlib==0.14.3
azure-cognitiveservices-speech==1.8.0
Babel==2.5.3
backports.shutil-get-terminal-size==1.0.0
beautifulsoup4==4.6.0
bitarray==0.8.1
bkcharts==0.2
blaze==0.11.3
bleach==2.1.2
bokeh==0.12.13
boto==2.48.0
boto3==1.14.4
botocore==1.17.4
Bottleneck==1.2.1
cachelib==0.1.1
cachetools==4.1.0
certifi==2018.1.18
cffi==1.14.0
chardet==3.0.4
click==6.7
cloudpickle==1.3.0
clyent==1.2.2
colorama==0.3.9
colorclass==2.2.0
coloredlogs==10.0
colorhash==1.0.2
comtypes==1.1.4
conda==4.6.14
conda-build==3.4.1
conda-verify==2.0.0
contextlib2==0.5.5
contextvars==2.4
cryptography==2.1.4
cycler==0.10.0
Cython==0.27.3
cytoolz==0.9.0
dask==0.16.1
datashape==0.5.4
decorator==4.4.2
distlib==0.3.0
distributed==1.20.2
dnspython==1.16.0
docopt==0.6.2
docutils==0.14
entrypoints==0.2.3
et-xmlfile==1.0.1
fastcache==1.0.2
fbmessenger==6.0.0
filelock==2.0.13
Flask==0.12.2
Flask-Cors==3.0.3
Flask-OAuthlib==0.9.5
Flask-RESTful==0.3.8
Flask-Session==0.3.2
future==0.18.2
gast==0.2.2
gevent==1.5.0
glob2==0.6
google-auth==1.17.2
google-auth-oauthlib==0.4.1
google-pasta==0.2.0
googletrans==2.4.0
greenlet==0.4.16
grpcio==1.29.0
h11==0.8.1
h2==3.2.0
h5py==2.7.1
heapdict==1.0.0
hpack==3.0.0
hstspreload==2020.5.27
html5lib==1.0.1
httplib2==0.18.1
httptools==0.1.1
httpx==0.9.3
humanfriendly==8.2
hyperframe==5.2.0
idna==2.6
idna-ssl==1.1.0
imageio==2.2.0
imagesize==0.7.1
immutables==0.14
importlib-metadata==1.6.0
importlib-resources==1.4.0
ipykernel==5.3.0
ipython==6.2.1
ipython-genutils==0.2.0
ipywidgets==7.1.1
isort==4.2.15
itsdangerous==0.24
jdcal==1.3
jedi==0.11.1
Jinja2==2.10
jmespath==0.10.0
joblib==0.15.1
jsonpickle==1.4.1
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==5.2.2
jupyter-console==5.2.0
jupyter-core==4.4.0
jupyterlab==0.31.4
jupyterlab-launcher==0.10.2
kafka-python==1.4.7
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
kiwisolver==1.2.0
lazy-object-proxy==1.3.1
llvmlite==0.21.0
locket==0.2.0
lxml==4.1.1
Markdown==3.2.2
MarkupSafe==1.0
matplotlib==3.2.1
mattermostwrapper==2.2
mccabe==0.6.1
menuinst==1.4.11
mistune==0.8.3
mpmath==1.0.0
msgpack-python==0.5.1
multidict==4.7.6
multipledispatch==0.4.9
navigator-updater==0.1.0
nbconvert==5.3.1
nbformat==4.4.0
networkx==2.4
nltk==3.2.5
nose==1.3.7
notebook==5.4.0
numba==0.36.2
numexpr==2.6.4
numpy==1.18.5
numpydoc==0.7.0
oauth2client==4.1.3
oauthlib==3.1.0
odo==0.5.1
olefile==0.45.1
openpyxl==2.4.10
opt-einsum==3.2.1
packaging==20.4
pandas==0.22.0
pandocfilters==1.4.2
parso==0.1.1
partd==0.3.8
path.py==10.5
pathlib2==2.3.0
patsy==0.5.0
pep8==1.7.1
pickleshare==0.7.4
pika==1.1.0
Pillow==5.0.0
pkginfo==1.4.1
pluggy==0.6.0
ply==3.10
prompt-toolkit==2.0.10
protobuf==3.12.2
psutil==5.4.3
psycopg2-binary==2.8.5
py==1.5.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.3.1
pycosat==0.6.3
pycparser==2.18
pycrypto==2.6.1
pycurl==7.43.0.1
pydot==1.4.1
pyflakes==1.6.0
Pygments==2.2.0
PyJWT==1.7.1
pykwalify==1.7.0
pylint==1.8.2
pymongo==3.8.0
pyodbc==4.0.22
pyOpenSSL==17.5.0
pyparsing==2.2.0
pyreadline==2.1
pyrsistent==0.16.0
PySocks==1.6.7
pytest==3.3.2
python-crfsuite==0.9.7
python-dateutil==2.8.1
python-dotenv==0.14.0
python-engineio==3.12.1
python-socketio==4.5.1
python-telegram-bot==12.7
pytz==2019.3
PyWavelets==0.5.2
pywin32==222
pywinpty==0.5
PyYAML==3.12
pyzmq==16.0.3
QtAwesome==0.4.4
qtconsole==4.3.1
QtPy==1.3.1
questionary==1.4.0
rasa==1.10.3
rasa-sdk==1.10.1
redis==3.5.3
requests==2.23.0
requests-oauthlib==1.3.0
requests-toolbelt==0.9.1
rfc3986==1.4.0
rocketchat-API==1.3.1
rope==0.10.7
rsa==4.6
ruamel-yaml==0.15.35
ruamel.yaml.clib==0.2.0
s3transfer==0.3.3
sanic==19.12.2
Sanic-Cors==0.10.0.post3
sanic-jwt==1.4.1
Sanic-Plugins-Framework==0.9.2
scikit-image==0.13.1
scikit-learn==0.22.2.post1
scipy==1.4.1
seaborn==0.8.1
Send2Trash==1.4.2
simplegeneric==0.8.1
singledispatch==3.4.0.3
six==1.11.0
sklearn-crfsuite==0.3.6
slackclient==2.7.1
sniffio==1.1.0
snowballstemmer==1.2.1
sortedcollections==0.5.3
sortedcontainers==1.5.9
Sphinx==1.6.6
sphinxcontrib-websupport==1.0.1
spyder==3.2.6
SQLAlchemy==1.3.17
statsmodels==0.8.0
sympy==1.1.1
tables==3.4.2
tabulate==0.8.7
tblib==1.3.2
tensorboard==2.1.1
tensorflow==2.1.1
tensorflow-addons==0.7.1
tensorflow-estimator==2.1.0
tensorflow-hub==0.8.0
tensorflow-probability==0.9.0
termcolor==1.1.0
terminado==0.8.1
terminaltables==3.1.0
testpath==0.3.1
toolz==0.9.0
tornado==6.0.4
tqdm==4.45.0
traitlets==4.3.2
twilio==6.26.3
typing==3.6.2
typing-extensions==3.7.4.2
tzlocal==2.1
ujson==2.0.3
unicodecsv==0.14.1
urllib3==1.22
wcwidth==0.2.4
webencodings==0.5.1
webexteamssdk==1.3
websockets==8.1
Werkzeug==0.14.1
widgetsnbextension==3.1.0
win-inet-pton==1.0.1
win-unicode-console==0.5
wincertstore==0.2
wrapt==1.12.1
xlrd==1.1.0
XlsxWriter==1.0.2
xlwings==0.11.5
xlwt==1.3.0
yarl==1.4.2
zict==0.1.3
zipp==3.1.0
